refactor(policy): move tenant enforcement to policy/engine
Spec 1.1 complete: Move tenant policy enforcement functions from
tenant-policy-helpers.ts to policy/engine/tenant-enforcement.ts

- Create policy/engine/tenant-enforcement.ts with:
  - TenantSettings interface
  - tenantModelAllowed
  - tenantProviderAllowed
  - filterTenantProviderRoutes
  - resolveExplicitTenantProviderId

- Update exports in policy/engine/index.ts
- Update imports in 7 dependent files:
  - model-routing-pipeline.ts
  - routes/chat.ts, images.ts, responses.ts, embeddings.ts, models.ts
  - tests/tenant-model-policy.test.ts

- Delete tenant-policy-helpers.ts

All 579 tests pass. No behavior changes.
